#425: an int expected against a long field, and eight bytes written into four
pthread_cond.c's lazy init compared the condvar signature with an `int'
expected value while the field is a `long'.  __atomic_compare_exchange_n takes
the expected value by ADDRESS and operates at the width of the OBJECT, so on
x86-64 it read eight bytes out of a four-byte slot and, on failure, wrote eight
back.

The comparison therefore failed for every thread, every thread took the losing
branch of the lazy init, and every one of them spun waiting for a signature
that nobody was left to publish.  pthread_test stopped on arm 2 of 23 with no
log to say why: a thread parked in a spin prints nothing.

i386 never saw it -- there `long' is four bytes, and the two widths coincide by
accident of the target.

pthread_once() in this same library had it right, with the same lazy-init
shape, which is what the fix is copied from.

+++ uros/src/lib/libpthreads/pthread_cond.c
@@ -48,7 +48,29 @@
 static int
 _pthread_cond_lazy_init(pthread_cond_t *cond)
 {
-	int expected = _PTHREAD_COND_SIG_init;
+	/*
+	 * 🔴 `long', because `sig' is a long -- and this was an `int' until
+	 * #425, which is the whole of that defect.
+	 *
+	 * __atomic_compare_exchange_n takes the expected value BY ADDRESS and
+	 * operates at the width of the object, so on a 64-bit target it reads
+	 * eight bytes from a four-byte stack slot and, when the compare fails,
+	 * writes eight bytes back into it.  The compare therefore fails for
+	 * every thread -- the upper four bytes are whatever the stack held --
+	 * every thread takes the losing branch, and every one of them spins
+	 * for a magic word nobody is left to publish.  pthread_test stopped
+	 * there, on the second of twenty-three arms.
+	 *
+	 * ⚠️ i386 never saw it: `long' is four bytes there, so the two widths
+	 * agree by accident of the target rather than by anything written.
+	 *
+	 * 🔑 gcc names it exactly -- "__atomic_compare_exchange_8 writing 8
+	 * bytes into a region of size 4 overflows the destination" -- and this
+	 * library is compiled -w, so it was thrown away.  pthread_once() has
+	 * the same lazy-init written correctly in another file, which is what
+	 * made the intent unambiguous.
+	 */
+	long expected = _PTHREAD_COND_SIG_init;
 	if (__atomic_compare_exchange_n(&cond->sig, &expected,
 					_PTHREAD_NO_SIG, 0,
 					__ATOMIC_ACQUIRE, __ATOMIC_ACQUIRE))
